[#3] Making repository more hexagonal

+++ src/facts/migrations/0001_facts_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS facts (
-  id bigint PRIMARY KEY,
+  id integer PRIMARY KEY,
   title varchar(64) NOT NULL CHECK (title <> ''),
   body varchar(2048) NOT NULL CHECK (body <> '')
 )
